Changed way for saving user in transaction during sign up via google token
