+++ service-api/pom.xml
@@ -12,7 +12,6 @@
     <properties>
         <jjwt.version>0.9.1</jjwt.version>
         <passay.version>1.6.0</passay.version>
-        <google.api-client.version>1.30.8</google.api-client.version>
     </properties>
 
     <artifactId>service-api</artifactId>
@@ -32,12 +31,6 @@
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.api-client</groupId>
-            <artifactId>google-api-client</artifactId>
-            <version>${google.api-client.version}</version>
-            <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>io.jsonwebtoken</groupId>
